feat(api): add separation file (#34)
